feat: add new folder for mmo poly
