+++ projects/mmo-finance-polygon/abi.json
@@ -1,0 +1,19 @@
+{
+  "inputs": [
+    {
+      "internalType": "address",
+      "name": "pool",
+      "type": "address"
+    }
+  ],
+  "name": "tvlOfPool",
+  "outputs": [
+    {
+      "internalType": "uint256",
+      "name": "tvl",
+      "type": "uint256"
+    }
+  ],
+  "stateMutability": "view",
+  "type": "function"
+}
